+++ src/App.js
@@ -1,11 +1,11 @@
 /* eslint-disable import/no-named-as-default */
 import { Route, Switch } from 'react-router-dom'
-import ChatApp from '../containers/chatApp'
-import Login from '../containers/login'
+import ChatApp from './containers/chatApp'
+import Login from './containers/login'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { hot } from 'react-hot-loader'
-import '../styles/styles.scss'
+import './styles/styles.scss'
 
 class App extends React.Component {
   render () {
