+++ src/index.js
@@ -4,7 +4,7 @@ import React from 'react'
 import { render } from 'react-dom'
 import { AppContainer } from 'react-hot-loader'
 import configureStore, { history } from './store/configureStore'
-import Root from './components/Root'
+import Root from './Root'
 import './styles/styles.scss' // Yep, that's right. You can import SASS/CSS files too! Webpack will run the associated loader and plug this into the page.
 import 'semantic-ui-css/semantic.min.css'
 require('./favicon.ico') // Tell webpack to load favicon.ico
@@ -18,8 +18,8 @@ render(
 )
 
 if (module.hot) {
-  module.hot.accept('./components/Root', () => {
-    const NewRoot = require('./components/Root').default
+  module.hot.accept('./Root', () => {
+    const NewRoot = require('./Root').default
     render(
       <AppContainer>
         <NewRoot store={store} history={history} />
